Refactor directory layout and components
